requirements.txt
pyserial
sqlalchemy
mysql-connector-python